requirements-test.txt
pytest-django
pytest-cov
coveralls
